Implement building of 'static' frameworks for the iPhoneOS and iPhoneSimulator targets. This is a semi-supported framework type that uses a static archive, rather than dylibs, for linking.
This allows the use of frameworks for link and header paths on iPhoneOS, where dylibs are not permitted due to code signing restrictions.
The iPhone platform SDKs do not include frameworks as a supported product type, so the framework is generated by using the built MacOSX framework as a canonical source.